Steam: fixed errors in x64
git-svn-id: http://svn.miranda-ng.org/main/trunk@8904 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/Steam/src/Steam/friend.h
@@ -44,7 +44,7 @@ namespace SteamWebApi
 			std::vector<Summary*> items;
 
 		public:
-			int GetItemCount() const { return items.size(); }
+			size_t GetItemCount() const { return items.size(); }
 			const Summary *GetAt(int idx) const { return items.at(idx); }
 		};
